requirements.txt
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
cattrs==23.2.3
cffi==1.16.0
cryptography==42.0.5
defusedxml==0.7.1
diff-match-patch==20230430
dj-user-login-history==1.0.5
Django==4.2.11
django-admin-honeypot==1.1.0
django-admin-honeypot-updated-2021==1.2.0
django-axes==6.4.0
django-easy-audit==1.3.5
django-honeypot==1.1.0
django-import-export==3.3.7
django-ipware==6.0.4
django-schema-graph==3.1.0
django-seed==0.3.1
et-xmlfile==1.1.0
Faker==24.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
psycopg2==2.9.9
pycparser==2.21
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ipware==2.0.2
PyYAML==6.0.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tablib==3.5.0
toposort==1.10
tzdata==2024.1
uv==0.1.17
xlrd==2.0.1
xlwt==1.3.0
